// src/openapi/codegen/models/FormDTO.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type FormDTO = {
    applicantAddress?: string;
    applicantCitizenship?: string;
    applicantGender?: string;
    applicantIc?: string;
    applicantMaritalStatus?: string;
    applicantMobileNumber?: string;
    applicantName?: string;
    applicantOccupation?: string;
    applicantPlaceOfBirth?: string;
    applicantPlaceOfBusiness?: string;
    applicantPostcode?: string;
    applicantResidencyPeriod?: string;
    applicantState?: string;
    applicationDate?: string;
    businessExp?: string;
    businessType?: string;
    capitalMoney?: number;
    endDate?: string;
    formDetails?: string;
    licenseNumber?: string;
    noOfDependents?: number;
    operationTime?: string;
    propertyStatus?: string;
    salesLocation?: string;
    salesType?: string;
    spouseAddress?: string;
    spouseIc?: string;
    spouseMobileNumber?: string;
    spouseName?: string;
    spouseOccupation?: string;
    startDate?: string;
    staticBusinessLocations?: string;
    vehicleNumber?: string;
    vehicleType?: string;
};